+++ src/ts/components/closebutton.ts
@@ -26,7 +26,7 @@ export class CloseButton extends Button<CloseButtonConfig> {
     }, this.config);
   }
 
-  configure(player: bitmovin.player.Player, uimanager: UIInstanceManager): void {
+  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
     let config = <CloseButtonConfig>this.getConfig();