+++ src/ts/components/controlbar.ts
@@ -25,7 +25,7 @@ export class ControlBar extends Container<ControlBarConfig> {
     }, <ControlBarConfig>this.config);
   }
 
-  configure(player: bitmovin.player.Player, uimanager: UIInstanceManager): void {
+  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
     // Counts how many components are hovered and block hiding of the control bar