+++ src/ts/components/titlebar.ts
@@ -33,7 +33,7 @@ export class TitleBar extends Container<TitleBarConfig> {
     }, <TitleBarConfig>this.config);
   }
 
-  configure(player: bitmovin.player.Player, uimanager: UIInstanceManager): void {
+  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
     let config = <TitleBarConfig>this.getConfig();